feat: add tiebreakers as option (#17)
